+++ CHANGELOG.md
@@ -10,8 +10,11 @@
 >
 > please add your unreleased change here.
 
-## TBD
+- [Feature] Add more send/recv actions profiling
 
+## 20240716
+
+- [SPU] 0.9.2b0 release
 - [Feature] Support jax.numpy.bitwise_count
 - [Bugfix] Fix jax.numpy.signbit wrong answer with very large input